# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NEXUS Playground API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Chat {
  id: number;
  userId: string;
  title: string;
  language: string;
  createdAt: string;
  updatedAt: string;
}

export interface Message {
  id: number;
  chatId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface ChatWithMessages {
  id: number;
  userId: string;
  title: string;
  language: string;
  createdAt: string;
  updatedAt: string;
  messages: Message[];
}

export interface CreateChatInput {
  title: string;
  language?: string;
}

export interface UpdateChatInput {
  title: string;
}

export interface AddMessageInput {
  role: string;
  content: string;
}