# lib/api-zod/src/generated/types/addMessageInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NEXUS Playground API
 * OpenAPI spec version: 0.1.0
 */

export interface AddMessageInput {
  role: string;
  content: string;
}