# lib/api-zod/src/generated/types/chatWithMessages.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NEXUS Playground API
 * OpenAPI spec version: 0.1.0
 */
import type { Message } from "./message";

export interface ChatWithMessages {
  id: number;
  userId: string;
  title: string;
  language: string;
  createdAt: Date;
  updatedAt: Date;
  messages: Message[];
}