# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NEXUS Playground API
 * OpenAPI spec version: 0.1.0
 */

export interface Message {
  id: number;
  chatId: number;
  role: string;
  content: string;
  createdAt: Date;
}